separate tests for karma and default framework
